+++ include/aero/fsi/MapLoad.h
@@ -26,15 +26,6 @@ linInterpVec(double* a, double* b, double interpFac, double* aInterpb)
     aInterpb[i] = a[i] + interpFac * (b[i] - a[i]);
 }
 
-std::array<double, 6>
-accumulateLoadsAndMoments(
-  const stk::mesh::BulkData& bulk,
-  const stk::mesh::PartVector& surface,
-  const sierra::nalu::VectorFieldType& modelCoords,
-  const sierra::nalu::VectorFieldType& meshDisp,
-  const sierra::nalu::GenericFieldType& tforceSCS,
-  std::array<double, 3> &center_of_mass);
-
 void mapTowerLoad(
   const stk::mesh::BulkData& bulk,
   const stk::mesh::PartVector& twrBndyParts,
